Cleanup SDK codebase
Consolidate Rally options in an interface Renamed rally.ts -> Rally.ts as it is a class

Refactored inline methods from constructor and made them private (and removed _)

Removed _stateChangeCallback as it is in options and validated in constructor

Renamed test file
